% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbhm_amplitude.R
\name{.sbhm_prewhiten}
\alias{.sbhm_prewhiten}
\title{Optionally prewhiten Y, regs (stacked), intercept and nuisance}
\usage{
.sbhm_prewhiten(Y, regs, Zint, Nuisance, prewhiten, X_other = NULL)
}
\description{
Optionally prewhiten Y, regs (stacked), intercept and nuisance
}
\keyword{internal}
