% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbhm_amplitude.R
\name{.sbhm_solve}
\alias{.sbhm_solve}
\title{Stable linear solve for multiple RHS with tiny ridge}
\usage{
.sbhm_solve(G, B, ridge = 0)
}
\description{
Stable linear solve for multiple RHS with tiny ridge
}
\keyword{internal}
