% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lss_cpp_optimized.R
\name{lss_cpp_optimized}
\alias{lss_cpp_optimized}
\title{A wrapper for the optimized C++ LSS implementation}
\usage{
lss_cpp_optimized(Y, bdes)
}
\arguments{
\item{Y}{the voxel by time data matrix}

\item{bdes}{the block design list created by \code{block_design}}
}
\value{
a matrix of beta estimates
}
\description{
A wrapper for the optimized C++ LSS implementation
}
\examples{
set.seed(1)
Y <- matrix(rnorm(16), 8, 2)
X_trials <- matrix(0, 8, 2)
X_trials[2:3, 1] <- 1
X_trials[5:6, 2] <- 1
bdes <- list(
  dmat_base = matrix(1, 8, 1),
  dmat_fixed = NULL,
  dmat_ran = X_trials
)
lss_cpp_optimized(Y, bdes)
}
