% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbhm_amplitude.R
\name{sbhm_amplitude_ls}
\alias{sbhm_amplitude_ls}
\title{Single-shape GLM amplitudes given matched coordinates (global LS)}
\usage{
sbhm_amplitude_ls(
  Y,
  sbhm,
  design_spec,
  alpha_hat,
  Nuisance = NULL,
  ridge = list(mode = "fractional", lambda = 0.02),
  prewhiten = NULL,
  return_se = FALSE
)
}
\description{
Single-shape GLM amplitudes given matched coordinates (global LS)
}
\keyword{internal}
