% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_hrf_recovery.R
\name{create_lwu_grid}
\alias{create_lwu_grid}
\title{Create LWU HRF Grid for OASIS Search}
\usage{
create_lwu_grid(
  tau_range = c(4, 8),
  sigma_range = c(1.5, 3.5),
  rho_range = c(0.1, 0.6),
  n_tau = 5,
  n_sigma = 3,
  n_rho = 3
)
}
\arguments{
\item{tau_range}{Range of tau values to test}

\item{sigma_range}{Range of sigma values to test}

\item{rho_range}{Range of rho values to test}

\item{n_tau}{Number of tau values in grid}

\item{n_sigma}{Number of sigma values in grid}

\item{n_rho}{Number of rho values in grid}
}
\value{
List of HRF models and their parameters
}
\description{
Generates a grid of LWU HRF models with varying parameters
}
