% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oasis_hrf_recovery.R
\name{fit_oasis_grid}
\alias{fit_oasis_grid}
\title{Fit OASIS with HRF Grid Search}
\usage{
fit_oasis_grid(Y, onsets, sframe, hrf_grid, ridge_x = 0.01, ridge_b = 0.01)
}
\arguments{
\item{Y}{Data matrix (time x voxels)}

\item{onsets}{Event onset times}

\item{sframe}{Sampling frame}

\item{hrf_grid}{List of HRF models to test}

\item{ridge_x}{Ridge parameter for design matrix}

\item{ridge_b}{Ridge parameter for aggregator}
}
\value{
List with best HRF index, parameters, and beta estimates
}
\description{
Fits OASIS models with different HRF parameters and selects best
}
