% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lss_cpp_optimized.R
\name{lss_cpp_optimized}
\alias{lss_cpp_optimized}
\title{A wrapper for the optimized C++ LSS implementation}
\usage{
lss_cpp_optimized(Y, bdes)
}
\arguments{
\item{Y}{the voxel by time data matrix}

\item{bdes}{the block design list created by \code{block_design}}
}
\value{
a matrix of beta estimates
}
\description{
A wrapper for the optimized C++ LSS implementation
}
